refactor: 도메인 로직을 domain 디렉토리 하위로 이동
