+++ __tests__/Car.test.js
@@ -1,4 +1,4 @@
-import Car from "../src/Car.js";
+import Car from "../src/domain/Car.js";
 
 test("랜덤 숫자가 4 이상일 경우 자동차는 움직인다.", () => {
   const car = new Car();
