+++ src/index.js
@@ -1,4 +1,4 @@
-import Race from "./Race.js";
+import Race from "./domain/Race.js";
 
 const race = new Race();
 await race.play();
